feat: Initial cut at functionality
This is a preliminary version of the neighborhood map.
In this version, I have loaded the map and a set of markers.
I also have a list-view of places. When you click on the places (either
in the list menu or the markers themselves), the program animates the
marker and displays a very basic infowindow for the selected place.