feat: Add functionality to push menu off canvas for smaller screens
Use media queries to change layout for smaller screens. On
device screens < 600px, the places menu is stored off canvas.
When the hamburger menu on the navbar is pressed, the menu will slide
on screen. A close button on the menu will move it back off screen again.

On wider screens, the places menu is permanently displayed alongside
the map.