fix: Clean up viewModel
Previously, I had not assigned a variable to the constructed viewModel
and simply created it as an unnamed input parameter to ko.bindings().
This had the unintended consequence that I could not access the
viewModel or any of its functions from the mapView. I got around this
by creating a separate VM that could interact with the map. Not good.

I corrected this by declaring a viewModel variable, so now I can access
the viewModel and its functions from the mapView.