feat: Add streetview image to infowindow
Changed the streetview call to use lat/lng coordinates, since this
gives a more accurate image than using the physical address.
Also added a streetview image to the infowindow.

At a later time, I will have to decide which of the two views I want
to keep, since I now display the image in both the DOM and the
infowindow. I will decide which is best based on which other API
I use.