refactor: Load API data on initial load for efficiency
Refactored code to load all the API data on initial load instead of
loading whenever a place is selected. This allows us to load only
once.

Also fixed a bug so that a closed infowindow will re-open, if the marker
is clicked again (even if the selectedPlace did not change).